First pass at DecorationStrategy
