+++ src/server/shell/decoration/CMakeLists.txt
@@ -9,7 +9,8 @@ set(
   basic_decoration.h    basic_decoration.cpp
   window.h              window.cpp
   input.h               input.cpp
-  renderer.h            renderer.cpp renderer_default.cpp
+  renderer.h            renderer.cpp
+  decoration_strategy.h decoration_strategy.cpp
 )
 
 add_library(
