+++ tests/unit-tests/shell/test_decoration_basic_decoration.cpp
@@ -187,7 +187,8 @@ struct DecorationBasicDecoration
             mt::fake_shared(buffer_allocator),
             mt::fake_shared(executor),
             mt::fake_shared(cursor_images),
-            mt::fake_shared(window_surface));
+            mt::fake_shared(window_surface),
+            msd::DecorationStrategy::default_decoration_strategy());
         executor.execute();
     }
